+++ catalyst/contrib/models/cv/segmentation/blocks/fpn.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+# @TODO: code formatting issue for 20.07 release
 import torch
 from torch import nn
 from torch.nn import functional as F