+++ catalyst/contrib/models/cv/segmentation/encoder/resnet.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+# @TODO: code formatting issue for 20.07 release
 from typing import List, Union
 from collections import OrderedDict
 from pathlib import Path