+++ catalyst/data/scripts/project_embeddings.py
@@ -1,3 +1,5 @@
+# flake8: noqa
+# @TODO: code formatting issue for 20.07 release
 import argparse
 import os
 from os import path